+++ src/app/static/src/app/store/dataExploration/dataExploration.ts
@@ -39,8 +39,6 @@ export interface DataExplorationState extends TranslatableState {
     errors: ErrorsState,
     currentUser: string,
     updatingLanguage: boolean,
-    errorReportError: Error | null
-    errorReportSuccess: boolean,
     dataExplorationMode: boolean
 }
 
@@ -49,8 +47,6 @@ export const initialDataExplorationState = (): DataExplorationState => {
         language: Language.en,
         version: currentHintVersion,
         updatingLanguage: false,
-        errorReportError: null,
-        errorReportSuccess: false,
         hintrVersion: initialHintrVersionState(),
         adr: initialADRState(),
         genericChart: initialGenericChartState(),