+++ src/app/static/src/app/store/root/actions.ts
@@ -2,10 +2,12 @@ import {ActionContext, ActionTree} from "vuex";
 import {RootState} from "../../root";
 import {StepDescription} from "../stepper/stepper";
 import {RootMutation} from "./mutations";
+import {ErrorsMutation} from "../errors/mutations";
 import {LanguageActions} from "../language/language";
 import {changeLanguage} from "../language/actions";
 import i18next from "i18next";
 import {api} from "../../apiService";
+import {ErrorReportManualDetails} from "../../types";
 import {VersionInfo} from "../../generated";
 import {currentHintVersion} from "../../hintVersion";
 import {LanguageMutation} from "../language/mutations";
@@ -15,23 +17,6 @@ export interface RootActions extends LanguageActions<RootState> {
     validate: (store: ActionContext<RootState, RootState>) => void;
     generateErrorReport: (store: ActionContext<RootState, RootState>,
                           payload: ErrorReportManualDetails) => void;
-}
-
-export interface ErrorReportManualDetails {
-    section: string,
-    description: string,
-    stepsToReproduce: string,
-    email: string
-}
-
-export interface ErrorReport extends ErrorReportManualDetails {
-    country: string,
-    projectName: string | undefined,
-    browserAgent: string,
-    timeStamp: string,
-    jobId: string,
-    versions: VersionInfo,
-    errors: Error[]
 }
 
 export const actions: ActionTree<RootState, RootState> & RootActions = {
@@ -112,18 +97,18 @@ export const actions: ActionTree<RootState, RootState> & RootActions = {
             timeStamp: new Date().toISOString(),
             jobId: rootState.modelRun.modelRunId || "no associated jobId",
             description: payload.description,
-            section: payload.section,
+            section: payload.section || "no associated section",
             stepsToReproduce: payload.stepsToReproduce,
             versions: {hint: currentHintVersion, ...rootState.hintrVersion.hintrVersion as VersionInfo},
             errors: getters.errors
         }
 
-        await api<RootMutation, RootMutation>(context)
-            .withSuccess(RootMutation.ErrorReportSuccess)
-            .withError(RootMutation.ErrorReportError)
+        await api<ErrorsMutation, ErrorsMutation>(context)
+            .withSuccess(`errors/${ErrorsMutation.ErrorReportSuccess}` as ErrorsMutation, true)
+            .withError(`errors/${ErrorsMutation.ErrorReportError}` as ErrorsMutation, true)
             .postAndReturn("error-report", data)
             .then(() => {
-                if (rootState.projects.currentProject && !rootState.errorReportError) {
+                if (rootState.projects.currentProject && !rootState.errors.errorReportError) {
                     dispatch("projects/cloneProject",
                         {
                             emails: ["naomi-support@imperial.ac.uk"],