+++ src/app/static/src/tests/components/dataExploration/dataExploration.test.ts
@@ -2,14 +2,16 @@ import {shallowMount} from "@vue/test-utils";
 import DataExploration from "../../../app/components/dataExploration/DataExploration.vue"
 import Vuex from "vuex";
 import {emptyState} from "../../../app/root";
-import {mockBaselineState, mockSurveyAndProgramState} from "../../mocks";
+import {
+    mockBaselineState,
+    mockSurveyAndProgramState
+} from "../../mocks";
 import {getters} from "../../../app/store/surveyAndProgram/getters";
 import registerTranslations from "../../../app/store/translations/registerTranslations";
 import {BaselineActions} from "../../../app/store/baseline/actions";
 import {SurveyAndProgramActions} from "../../../app/store/surveyAndProgram/actions";
 import {BaselineState} from "../../../app/store/baseline/baseline";
 import {DataType, SurveyAndProgramState} from "../../../app/store/surveyAndProgram/surveyAndProgram";
-import {expectTranslatedWithStoreType} from "../../testHelpers";
 
 describe(`data exploration component`, () => {
     let actions: jest.Mocked<BaselineActions>;
@@ -65,25 +67,35 @@ describe(`data exploration component`, () => {
         const wrapper = shallowMount(DataExploration, {store});
         expect(wrapper.find("adr-integration-stub").exists()).toBe(true)
         expect(wrapper.find("upload-inputs-stub").exists()).toBe(true)
-        expect(wrapper.findAll("button").length).toBe(2)
+        expect(wrapper.findAll("stepper-navigation-stub").exists()).toBe(true)
     })
 
-    it(`can translate button texts`, () => {
+    it(`disables back navigation when on upload step`, () => {
         const store = createSut()
         const wrapper = shallowMount(DataExploration, {store});
-        expect(wrapper.findAll("button").length).toBe(2)
-        const buttons = wrapper.findAll("button")
-        expectTranslatedWithStoreType(buttons.at(0), "back", "retour", "voltar", store)
-        expectTranslatedWithStoreType(buttons.at(1), "next", "suivant", "próximo", store)
+        expect(wrapper.find("stepper-navigation-stub").props("backDisabled")).toBe(true)
+        expect(wrapper.find("stepper-navigation-stub").props("nextDisabled")).toBe(false)
     })
 
-    it(`next button can navigate to reviewInputs`, () => {
+    it(`disables continue navigation when on review step`, () => {
+        const store = createSut()
+        const wrapper = shallowMount(DataExploration, {
+            store,
+            data() {
+                return {step: 2}
+            }
+        });
+
+        expect(wrapper.find("stepper-navigation-stub").props("backDisabled")).toBe(false)
+        expect(wrapper.find("stepper-navigation-stub").props("nextDisabled")).toBe(true)
+    })
+
+    it(`continue button can navigate to reviewInputs`, () => {
         const store = createSut()
         const wrapper = shallowMount(DataExploration, {store});
-        expect(wrapper.findAll("button").length).toBe(2)
         expect(wrapper.find("review-inputs-stub").exists()).toBe(false)
-        const buttons = wrapper.findAll("button")
-        buttons.at(1).trigger("click")
+        const vm = wrapper.vm as any
+        vm.next()
         expect(wrapper.find("review-inputs-stub").exists()).toBe(true)
     })
 
@@ -95,10 +107,10 @@ describe(`data exploration component`, () => {
                 return {step: 2}
             }
         });
-        expect(wrapper.findAll("button").length).toBe(2)
+
         expect(wrapper.find("upload-inputs-stub").exists()).toBe(false)
-        const buttons = wrapper.findAll("button")
-        buttons.at(0).trigger("click")
+        const vm = wrapper.vm as any
+        vm.back()
         expect(wrapper.find("upload-inputs-stub").exists()).toBe(true)
     })