+++ src/app/static/src/tests/components/errorReport.test.ts
@@ -3,7 +3,7 @@ import {shallowMount, mount, createLocalVue} from '@vue/test-utils';
 import ErrorReport from "../../app/components/ErrorReport.vue";
 import Modal from "../../app/components/Modal.vue";
 import Vuex from "vuex";
-import {mockProjectsState, mockRootState, mockStepperState} from "../mocks";
+import {mockErrorsState, mockProjectsState, mockRootState, mockStepperState} from "../mocks";
 import registerTranslations from "../../app/store/translations/registerTranslations";
 import {StepperState} from "../../app/store/stepper/stepper";
 import {ProjectsState} from "../../app/store/projects/projects";
@@ -12,6 +12,7 @@ import VueRouter from "vue-router";
 import {Language} from "../../app/store/translations/locales";
 import ErrorAlert from "../../app/components/ErrorAlert.vue";
 import {RootState} from "../../app/root";
+import { ErrorsState } from "../../app/store/errors/errors";
 
 describe("Error report component", () => {
 
@@ -20,7 +21,8 @@ describe("Error report component", () => {
     const createStore = (stepperState: Partial<StepperState> = {},
                          projectsState: Partial<ProjectsState> = {},
                          rootState: Partial<RootState> = {},
-                         isGuest = false) => {
+                         isGuest = false,
+                         errorsState: Partial<ErrorsState> ={}) => {
         const store = new Vuex.Store({
             state: mockRootState(rootState),
             modules: {
@@ -31,7 +33,11 @@ describe("Error report component", () => {
                 projects: {
                     namespaced: true,
                     state: mockProjectsState(projectsState)
-                }
+                },
+                errors: {
+                    namespaced: true,
+                    state: mockErrorsState(errorsState)
+                },
             },
             actions: {
                 generateErrorReport
@@ -131,7 +137,7 @@ describe("Error report component", () => {
         expectTranslated(labels.at(4),
             "Steps to reproduce",
             "Étapes à reproduire",
-            "Reportar problemas",
+            "Passos para reproduzir",
             store)
 
         const mutedText = wrapper.findAll("div.small.text-muted")
@@ -591,7 +597,7 @@ describe("Error report component", () => {
             propsData: {
                 open: true
             },
-            store: createStore({}, {}, {errorReportError})
+            store: createStore({}, {}, {}, false, {errorReportError})
         });
         wrapper.setData({showFeedback: true});
         expectTranslated(wrapper.find("#report-error"),